Refine handling of 'card' parameter to support generic integrations
When setting up a paypal rest Authorize request there are a few possibilities

we have a payment token
we have a credit card number & cvv & expiry
we have neither of the above and should be redirected to paypal for us to log in and authorize the payment.
Currently the code understands the difference as

cardReference is present
'card' object is present
neither of the above.
However, the card object also holds other information - email, address, phone etc. Even if these are not used by Paypal Rest the calling code should not need to know that it can't set these under a specific combination of circumstances.

This fix looks more deeply into the passed card parameter for the presence of the actual card fields - ie.card number, expiry - and otherwise concludes this is not a card-present transations and allows the redirect to proceed