Return token id if other possible values not present in getTransactionReference.
When processing a request with no card present a token is retrieved from
paypal which is the used by a paypal browser script to open a paypal popup to approve the token.

It's a bit arguable whether this is best retrieved via getTransactionReference or
getCardReference. cardReference makes sense in that the token is used in place of a card
but it's also the only reference we have for the transaction at this point.

Related :
https://github.com/thephpleague/omnipay-paypal/issues/219

https://github.com/thephpleague/omnipay-paypal/issues/215